Código main na view
